apps/dictation_for_dental_hygienists/pyproject.toml
[project]
name = "dictation-for-dental-hygienists"
version = "0.1.0"
description = "Voice dictation app for dental hygienists to record periodontal examination findings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.47.0",
    "dash>=3.3.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
